refactor FileTree
